pyproject.toml
[project]
name = "kickstarter-china-tracker"
version = "0.1.0"
description = "Track China-background consumer-hardware projects on Kickstarter (prelaunch + live)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
  "httpx[http2]>=0.27",
  "selectolax>=0.3.21",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8", "ruff>=0.5"]

[project.scripts]
ks-tracker = "scraper.run:run"

[tool.ruff]
line-length = 100
